Dear Dr. Harris, I made test cases. Love, Price

+++ bgp/input/ASTIf.c
@@ -1,4 +1,4 @@
-float main()
+int main()
 {
     if(1 == 2)
     {
